More work on .MIF voxel loading.
Some of the floor, wall, and platform voxel IDs are being loaded now and
can be written into the voxel grid. The textures are all wrong, though,
since the .INF file isn't being loaded yet (and it's just using my test
city textures), so that is probably something I'll look into next.

+++ OpenTESArena/src/Assets/MIFFile.h
@@ -66,9 +66,9 @@ public:
 	~MIFFile();
 
 	// Identifiers for various chasms in Arena's voxel data.
-	static const uint16_t DRY_CHASM;
-	static const uint16_t WET_CHASM;
-	static const uint16_t LAVA_CHASM;
+	static const uint8_t DRY_CHASM;
+	static const uint8_t WET_CHASM;
+	static const uint8_t LAVA_CHASM;
 
 	// Gets the dimensions of the map. They are constant for all levels in a map.
 	int getWidth() const;
